+++ test/repl.jl
@@ -684,7 +684,7 @@ end
 let io = IOBuffer()
     Base.display_error(io,
         try
-            [][trues(6000)]
+            [][BitVector(true, 6000)]
         catch e
             e
         end, [])